requirements.txt
django
django-quill-editor
psycopg2-binary
gunicorn
Pillow
django-heroku